feat(deployments): inline-edit deployment name from the detail pane header
