+++ src/renderer/src/deployments.jsx
@@ -1,6 +1,6 @@
 import L from 'leaflet'
 import 'leaflet/dist/leaflet.css'
-import { Camera, Check, ChevronDown, ChevronRight, MapPin, Pencil, X } from 'lucide-react'
+import { Camera, ChevronDown, ChevronRight, MapPin, X } from 'lucide-react'
 import { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import ReactDOMServer from 'react-dom/server'
 import { LayersControl, MapContainer, Marker, TileLayer, useMap, useMapEvents } from 'react-leaflet'
@@ -13,6 +13,7 @@ import { useImportStatus } from '@renderer/hooks/import'
 import SkeletonDeploymentsList from './ui/SkeletonDeploymentsList'
 import { resolveSelectedDeployment, withDeploymentParam } from './deployments/urlState'
 import DeploymentDetailPane from './deployments/DeploymentDetailPane'
+import EditableLocationName from './deployments/EditableLocationName'
 
 // Fix the default marker icon issue in react-leaflet
 // This is needed because the CSS assets are not properly loaded
@@ -397,137 +398,6 @@ function LocationMap({
     </div>
   )
 }
-
-// Editable location name component with inline editing
-const EditableLocationName = memo(function EditableLocationName({
-  locationID,
-  locationName,
-  isSelected,
-  onRename
-}) {
-  const [isEditing, setIsEditing] = useState(false)
-  const [editValue, setEditValue] = useState('')
-  const [isSaving, setIsSaving] = useState(false)
-  const inputRef = useRef(null)
-
-  const displayName = locationName || locationID || 'Unnamed Location'
-
-  const startEditing = useCallback(
-    (e) => {
-      e.stopPropagation()
-      setEditValue(displayName)
-      setIsEditing(true)
-    },
-    [displayName]
-  )
-
-  const cancelEditing = useCallback(() => {
-    setIsEditing(false)
-    setEditValue('')
-  }, [])
-
-  const saveEdit = useCallback(async () => {
-    const trimmed = editValue.trim()
-    // Cancel if empty or unchanged
-    if (!trimmed || trimmed === displayName) {
-      cancelEditing()
-      return
-    }
-
-    setIsSaving(true)
-    try {
-      await onRename(locationID, trimmed)
-      setIsEditing(false)
-    } catch (error) {
-      console.error('Error renaming location:', error)
-      // Keep edit mode open for retry
-    } finally {
-      setIsSaving(false)
-    }
-  }, [editValue, displayName, locationID, onRename, cancelEditing])
-
-  const handleKeyDown = useCallback(
-    (e) => {
-      if (e.key === 'Enter') {
-        e.preventDefault()
-        saveEdit()
-      } else if (e.key === 'Escape') {
-        e.preventDefault()
-        cancelEditing()
-      }
-    },
-    [saveEdit, cancelEditing]
-  )
-
-  const handleBlur = useCallback(() => {
-    if (!isSaving) {
-      saveEdit()
-    }
-  }, [isSaving, saveEdit])
-
-  // Focus and select input when entering edit mode
-  useEffect(() => {
-    if (isEditing && inputRef.current) {
-      inputRef.current.focus()
-      inputRef.current.select()
-    }
-  }, [isEditing])
-
-  if (isEditing) {
-    return (
-      <div className="flex items-center gap-1" onClick={(e) => e.stopPropagation()}>
-        <input
-          ref={inputRef}
-          type="text"
-          value={editValue}
-          onChange={(e) => setEditValue(e.target.value)}
-          onKeyDown={handleKeyDown}
-          onBlur={handleBlur}
-          maxLength={100}
-          disabled={isSaving}
-          className="text-sm border border-blue-400 rounded px-1.5 py-0.5 w-48 focus:outline-none focus:ring-1 focus:ring-blue-500"
-        />
-        <button
-          onClick={saveEdit}
-          disabled={isSaving}
-          className="p-0.5 hover:bg-green-100 rounded text-green-600"
-          title="Save (Enter)"
-        >
-          <Check size={14} />
-        </button>
-        <button
-          onClick={cancelEditing}
-          disabled={isSaving}
-          className="p-0.5 hover:bg-red-100 rounded text-red-600"
-          title="Cancel (Esc)"
-        >
-          <X size={14} />
-        </button>
-      </div>
-    )
-  }
-
-  return (
-    <div className="group flex items-center gap-1 min-w-0">
-      <span
-        onClick={startEditing}
-        className={`cursor-pointer text-sm truncate min-w-0 ${
-          isSelected ? 'font-semibold text-blue-700' : 'text-gray-700'
-        }`}
-        title={`${displayName} (click to rename)`}
-      >
-        {displayName}
-      </span>
-      <button
-        onClick={startEditing}
-        className="p-0.5 opacity-0 group-hover:opacity-100 hover:bg-gray-200 rounded text-gray-500 transition-opacity flex-shrink-0"
-        title="Rename"
-      >
-        <Pencil size={12} />
-      </button>
-    </div>
-  )
-})
 
 // Memoized deployment row component
 const DeploymentRow = memo(function DeploymentRow({
@@ -1370,6 +1240,7 @@ export default function Deployments({ studyId }) {
               <DeploymentDetailPane
                 deployment={selectedLocation}
                 onClose={() => setSelectedLocation(null)}
+                onRenameLocation={onRenameLocation}
               />
             </Panel>
           </>
