+++ src/renderer/src/deployments/DeploymentDetailPane.jsx
@@ -1,25 +1,30 @@
 import { X } from 'lucide-react'
 import DeploymentMediaGallery from '../media/DeploymentMediaGallery'
+import EditableLocationName from './EditableLocationName'
 
 /**
  * Bottom-pane container for the Deployments tab. Mounted only when a
- * deployment is selected. Header shows the deployment name and a close
- * button. Body for V1 contains DeploymentMediaGallery; later additions
- * (timeline graph, camera-days, species at location) slot in as siblings
- * inside the body.
+ * deployment is selected. Header shows the inline-editable deployment
+ * name and a close button. Body for V1 contains DeploymentMediaGallery;
+ * later additions (timeline graph, camera-days, species at location)
+ * slot in as siblings inside the body.
  */
-export default function DeploymentDetailPane({ deployment, onClose }) {
-  const title = deployment.locationName || deployment.locationID || deployment.deploymentID
-
+export default function DeploymentDetailPane({ deployment, onClose, onRenameLocation }) {
   return (
     <div className="flex flex-col h-full bg-white border-t border-gray-200 min-h-0">
-      <div className="flex items-center justify-between px-4 py-2 border-b border-gray-200 flex-shrink-0">
-        <h2 className="text-sm font-semibold text-gray-700 truncate" title={title}>
-          {title} — media
-        </h2>
+      <div className="flex items-center justify-between px-4 py-2 border-b border-gray-200 flex-shrink-0 gap-2">
+        <div className="flex items-center gap-2 min-w-0">
+          <EditableLocationName
+            locationID={deployment.locationID || deployment.deploymentID}
+            locationName={deployment.locationName}
+            isSelected={true}
+            onRename={onRenameLocation}
+          />
+          <span className="text-sm text-gray-400 flex-shrink-0">— media</span>
+        </div>
         <button
           onClick={onClose}
-          className="p-1 hover:bg-gray-100 rounded text-gray-500 hover:text-gray-700"
+          className="p-1 hover:bg-gray-100 rounded text-gray-500 hover:text-gray-700 flex-shrink-0"
           title="Close (Esc)"
           aria-label="Close media pane"
         >
